Reorganize service files into a separate subdirectory
