+++ src/index.js
@@ -2,11 +2,11 @@ import $ from 'jquery';
 import 'bootstrap';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './css/styles.css';
-import Geocode  from './geocode.js';
+import Geocode  from './services/geocode.js';
 import GeocodeErrorHandling from './geocode-error-handling.js';
-import NearbyService from './ebird-service.js';
+import NearbyService from './services/ebird-service.js';
 import Validation from './validation.js';
-// import BirdSoundsService from './bird-sounds-service.js';
+// import BirdSoundsService from './services/bird-sounds-service.js';
 // import DataParsing from './data-parsing.js';
 
 // function listNearbyBirds(response){
